don't push to history, dark mode theme improvements
search results are now never pushed to history, as one would not expect the search box to pop up when navigating. Also fixed unintentional re-calculation of results when the focus was taken out of the search box. Additionally, the fetch and clear results delay was reduced to provide a slightly better experience. Fixed contrast issues with dark mode and followed Material guidelines.